restructure to create python module namespace
